+++ assets/javascripts/kitten/components/form/field/index.js
@@ -2,6 +2,7 @@ import { Component } from 'react'
 import { FieldLabel } from './components/label'
 import { FieldInput } from './components/input'
 import { FieldError } from './components/error'
+import { FieldHelp } from './components/help'
 import { FieldRadioSet } from './components/radio-set'
 import { FieldRadioButtonSet } from './components/radio-button-set'
 import { FieldCheckbox } from './components/checkbox'
@@ -15,6 +16,7 @@ export class Field extends Component {
   static RadioButtonSet = FieldRadioButtonSet
   static Checkbox = FieldCheckbox
   static ErrorMessage = FieldError
+  static Help = FieldHelp
   static Password = FieldPassword
   static Autocomplete = FieldAutocomplete